requirements.txt
# Core dependencies
MetaTrader5>=5.0.45
python-dotenv>=1.0.0
pydantic>=2.0.0

# API & Backend
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
websockets>=11.0
requests>=2.31.0

# Data analysis and manipulation
pandas>=2.0.0
numpy>=1.24.0
